requirements.txt
python_telegram_bot
requests
